TVIST1-457: Added permission checks to controller actions

+++ src/Controller/AgendaController.php
@@ -30,6 +30,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Uid\UuidV4;
 
 /**
@@ -149,6 +150,10 @@ class AgendaController extends AbstractController
      */
     public function create(MunicipalityHelper $municipalityHelper, Request $request): Response
     {
+        if (!($this->isGranted('ROLE_CASEWORKER') || $this->isGranted('ROLE_ADMINISTRATION'))) {
+            throw new AccessDeniedException();
+        }
+
         $agenda = new Agenda();
 
         $form = $this->createForm(AgendaNewType::class, $agenda, [
@@ -180,6 +185,8 @@ class AgendaController extends AbstractController
      */
     public function delete(Agenda $agenda, Request $request): Response
     {
+        $this->denyAccessUnlessGranted('delete', $agenda);
+
         // Check that CSRF token is valid
         if ($this->isCsrfTokenValid('delete'.$agenda->getId(), $request->request->get('_token')) && !$agenda->isFinished()) {
             $this->entityManager->remove($agenda);
@@ -226,6 +233,8 @@ class AgendaController extends AbstractController
      */
     public function edit(Agenda $agenda, Request $request): ?Response
     {
+        $this->denyAccessUnlessGranted('edit', $agenda);
+
         $agendaOptions = $agenda->isFinished() ? ['disabled' => true] : [];
 
         $form = $this->createForm(AgendaEditType::class, $agenda, $agendaOptions);
@@ -254,6 +263,8 @@ class AgendaController extends AbstractController
      */
     public function addBoardMember(Agenda $agenda, BoardMemberRepository $memberRepository, Request $request): Response
     {
+        $this->denyAccessUnlessGranted('edit', $agenda);
+
         $availableBoardMembers = $memberRepository->getAvailableBoardMembersByAgenda($agenda);
 
         $form = $this->createForm(AgendaAddBoardMemberType::class, [], [
@@ -292,6 +303,8 @@ class AgendaController extends AbstractController
      */
     public function removeBoardMember(Agenda $agenda, BoardMember $boardMember, Request $request): Response
     {
+        $this->denyAccessUnlessGranted('edit', $agenda);
+
         // Check that CSRF token is valid
         if ($this->isCsrfTokenValid('remove'.$boardMember->getId(), $request->request->get('_token')) && !$agenda->isFinished()) {
             $agenda->removeBoardmember($boardMember);
@@ -306,6 +319,8 @@ class AgendaController extends AbstractController
      */
     public function protocol(Agenda $agenda, Request $request): Response
     {
+        $this->denyAccessUnlessGranted('employee', $agenda);
+
         $agendaProtocol = $agenda->getProtocol() ?? new AgendaProtocol();
 
         $agendaOptions = $this->agendaHelper->getFormOptionsForAgenda($agenda);
@@ -341,6 +356,8 @@ class AgendaController extends AbstractController
      */
     public function broadcastAgenda(Agenda $agenda, Request $request): Response
     {
+        $this->denyAccessUnlessGranted('employee', $agenda);
+
         $agendaOptions = $this->agendaHelper->getFormOptionsForAgenda($agenda);
 
         $form = $this->createForm(AgendaBroadcastType::class, null, $agendaOptions);
@@ -367,6 +384,8 @@ class AgendaController extends AbstractController
      */
     public function publishAgenda(Agenda $agenda): Response
     {
+        $this->denyAccessUnlessGranted('employee', $agenda);
+
         $agenda->setIsPublished(true);
         $this->entityManager->flush();
